refactor: change folders names and update imports
