+++ pricing-react/src/components/UI/ButtonsWrapper/ButtonsWrapper.jsx
@@ -1,4 +1,4 @@
-import Button from "../../UI/Button/Button";
+import Button from "../Button/Button";
 
 import styles from "./ButtonsWrapper.module.scss";
