+++ pricing-react/src/components/pricing/PricingCardsContainer/PricingCard/Features/Features.module.scss
