+++ pricing-react/src/components/pricing/PricingToggle/PricingToggle.module.scss
